webservice eliminar preenvios
eliminacion de las cajas y tarimas que pertenecen a ese envio.

+++ webservice/eliminarEnvios.php
@@ -12,16 +12,21 @@
 		case 1://productor
 		break;
 		case 2://empaque
-				$query = "DELETE FROM envios_empaque WHERE id_envio = $id_envio";
-				$r = mysql_query($query);
 
-				if($r){
-					$datos_usuario = "Bien*Pre-Envio eliminado correctamente";
-					mysql_query("UPDATE camiones_empaque set estado = 0 where id_camion = $carro");
-				}
-				else
-					$datos_usuario = "Error*No se pudo eliminar el envio $id_envio";	
+				$query = "DELETE FROM distribuidor_cajas_envio WHERE id_envio_fk = $id_envio";
+				$result = mysql_query($query);
+				if($result){
+					$query = "DELETE FROM envios_empaque WHERE id_envio = $id_envio";
+					$r = mysql_query($query);
 
+					if($r){
+						$datos_usuario = "Bien*Pre-Envio eliminado correctamente";
+						mysql_query("UPDATE camiones_empaque set estado = 0 where id_camion = $carro");
+					}
+					else
+						$datos_usuario = "Error*No se pudo eliminar el envio $id_envio";
+				}else	
+					$datos_usuario = "Error*No es posible eliminar las cajas y tarimas del envio $id_envio";
 		break;
 		case 3://distribuidor
 		break;
